Move point-in-poly functions to separate file and rename
