+++ src/apps/testapps/testPolyfill.c
@@ -318,41 +318,4 @@ TEST(polyfillPentagon) {
     free(hexagons);
 }
 
-TEST(_pointInPolyContainsLoop) {
-    GeoCoord somewhere = {1, 2};
-
-    BBox bbox;
-    bboxFromGeofence(&sfGeofence, &bbox);
-
-    t_assert(_pointInPolyContainsLoop(&sfGeofence, &bbox, &sfVerts[0]) == false,
-             "contains exact");
-    t_assert(_pointInPolyContainsLoop(&sfGeofence, &bbox, &sfVerts[4]) == true,
-             "contains exact4");
-    t_assert(_pointInPolyContainsLoop(&sfGeofence, &bbox, &somewhere) == false,
-             "contains somewhere else");
-}
-
-TEST(_pointInPolyContainsLoopTransmeridian) {
-    GeoCoord eastPoint = {0.001, -M_PI + 0.001};
-    GeoCoord eastPointOutside = {0.001, -M_PI + 0.1};
-    GeoCoord westPoint = {0.001, M_PI - 0.001};
-    GeoCoord westPointOutside = {0.001, M_PI - 0.1};
-
-    BBox bbox;
-    bboxFromGeofence(&transMeridianGeofence, &bbox);
-
-    t_assert(_pointInPolyContainsLoop(&transMeridianGeofence, &bbox,
-                                      &westPoint) == true,
-             "contains point to the west of the antimeridian");
-    t_assert(_pointInPolyContainsLoop(&transMeridianGeofence, &bbox,
-                                      &eastPoint) == true,
-             "contains point to the east of the antimeridian");
-    t_assert(_pointInPolyContainsLoop(&transMeridianGeofence, &bbox,
-                                      &westPointOutside) == false,
-             "does not contain outside point to the west of the antimeridian");
-    t_assert(_pointInPolyContainsLoop(&transMeridianGeofence, &bbox,
-                                      &eastPointOutside) == false,
-             "does not contain outside point to the east of the antimeridian");
-}
-
 END_TESTS();
